move topology control into its own crate
