+++ common/topology-control/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "nym-topology-control"
+version = "1.1.15"
+authors = ["Dave Hrycyszyn <futurechimp@users.noreply.github.com>"]
+edition = "2021"
+rust-version = "1.66"
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = { workspace = true }
+futures = { workspace = true }
+log = { workspace = true }
+rand = { version = "0.7.3", features = ["wasm-bindgen"] }
+serde = { workspace = true, features = ["derive"] }
+tap = "1.0.1"
+url = { workspace = true, features = ["serde"] }
+tokio = { workspace = true, features = ["macros"]}
+
+# internal
+
+nym-explorer-client = { path = "../../explorer-api/explorer-client" }
+nym-sphinx = { path = "../nymsphinx" }
+nym-topology = { path = "../topology", features = ["serializable"] }
+nym-validator-client = { path = "../client-libs/validator-client", default-features = false }
+nym-task = { path = "../task" }
+nym-network-defaults = { path = "../network-defaults" }
+
+
+[target."cfg(not(target_arch = \"wasm32\"))".dependencies.tokio-stream]
+version = "0.1.11"
+features = ["time"]
+
