+++ gateway/src/node/mod.rs
@@ -24,10 +24,6 @@ use anyhow::bail;
 use dashmap::DashMap;
 use futures::channel::{mpsc, oneshot};
 use log::*;
-use nym_client_core::client::topology_control::accessor::TopologyAccessor;
-use nym_client_core::client::topology_control::nym_api_provider::NymApiTopologyProvider;
-use nym_client_core::client::topology_control::TopologyRefresher;
-use nym_client_core::client::topology_control::TopologyRefresherConfig;
 use nym_crypto::asymmetric::{encryption, identity};
 use nym_mixnet_client::forwarder::{MixForwardingSender, PacketForwarder};
 use nym_network_defaults::NymNetworkDetails;
@@ -36,6 +32,10 @@ use nym_node::wireguard::types::GatewayClientRegistry;
 use nym_statistics_common::collector::StatisticsSender;
 use nym_task::{TaskClient, TaskManager};
 use nym_topology::provider_trait::TopologyProvider;
+use nym_topology_control::accessor::TopologyAccessor;
+use nym_topology_control::nym_api_provider::NymApiTopologyProvider;
+use nym_topology_control::TopologyRefresher;
+use nym_topology_control::TopologyRefresherConfig;
 use nym_validator_client::{nyxd, DirectSigningHttpRpcNyxdClient};
 use rand::seq::SliceRandom;
 use rand::thread_rng;
