+++ sdk/rust/nym-sdk/Cargo.toml
@@ -20,6 +20,7 @@ nym-network-defaults = { path = "../../../common/network-defaults" }
 nym-sphinx = { path = "../../../common/nymsphinx" }
 nym-task = { path = "../../../common/task" }
 nym-topology = { path = "../../../common/topology" }
+nym-topology-control = { path = "../../../common/topology-control" }
 nym-socks5-client-core = { path = "../../../common/socks5-client-core" }
 nym-validator-client = { path = "../../../common/client-libs/validator-client", features = ["http-client"] }
 nym-socks5-requests = { path = "../../../common/socks5/requests" }
